[#2][CR-2] rename file dto to match class name
