+++ src/routes/api/video/dto/index.ts
@@ -3,5 +3,5 @@ export { GetVideosRequest } from '@app-routes/api/video/dto/GetVideosRequest.dto
 export { VideoResponse } from '@app-routes/api/video/dto/VideoResponse.dto';
 export { GetVideosResponse } from '@app-routes/api/video/dto/GetVideosResponse.dto';
 export { AnimeEpisodeInfo } from '@app-routes/api/video/dto/AnimeEpisodeInfo.dto';
-export { SearchVideosRequest } from '@app-routes/api/video/dto/GetVideosByUploaderRequest.dto';
+export { SearchVideosRequest } from '@app-routes/api/video/dto/SearchVideosRequest.dto';
 export { GetVideosInfoRequest } from '@app-routes/api/video/dto/GetVideosInfoRequest.dto';
